chore: reduce number of client types
