+++ lib/index.ts
@@ -5,4 +5,3 @@ export * from './utils/normalize-select'
 export * from './utils/resolve-circular'
 
 export * from './types'
-export type { ContentfulClientApi } from './create-contentful-api'
