+++ lib/types/index.ts
@@ -1,5 +1,6 @@
 export * from './asset'
 export * from './asset-key'
+export { ContentfulClientApi } from './client'
 export * from './collection'
 export * from './content-type'
 export * from './entry'
