+++ test/types/client/createClient.test-d.ts
@@ -1,16 +1,17 @@
 import { expectNotAssignable, expectType } from 'tsd'
 
-import { createClient } from '../../../lib'
-import { Client } from '../../../lib/create-contentful-api'
+import { ContentfulClientApi, createClient } from '../../../lib'
 
 const CLIENT_OPTIONS = {
   accessToken: 'accessToken',
   space: 'spaceId',
 }
 
-expectType<Client<undefined>>(createClient(CLIENT_OPTIONS))
+expectType<ContentfulClientApi<undefined>>(createClient(CLIENT_OPTIONS))
 
-expectType<Client<'WITHOUT_LINK_RESOLUTION'>>(createClient(CLIENT_OPTIONS).withoutLinkResolution)
+expectType<ContentfulClientApi<'WITHOUT_LINK_RESOLUTION'>>(
+  createClient(CLIENT_OPTIONS).withoutLinkResolution
+)
 expectNotAssignable<{ withoutLinkResolution: any }>(
   createClient(CLIENT_OPTIONS).withoutLinkResolution
 )
@@ -18,7 +19,7 @@ expectNotAssignable<{ withoutUnresolvableLinks: any }>(
   createClient(CLIENT_OPTIONS).withoutLinkResolution
 )
 
-expectType<Client<'WITHOUT_LINK_RESOLUTION' | 'WITH_ALL_LOCALES'>>(
+expectType<ContentfulClientApi<'WITHOUT_LINK_RESOLUTION' | 'WITH_ALL_LOCALES'>>(
   createClient(CLIENT_OPTIONS).withoutLinkResolution.withAllLocales
 )
 expectNotAssignable<{ withoutLinkResolution: any }>(
@@ -31,7 +32,7 @@ expectNotAssignable<{ withoutLinkResolution: any }>(
   createClient(CLIENT_OPTIONS).withoutLinkResolution.withAllLocales
 )
 
-expectType<Client<'WITHOUT_UNRESOLVABLE_LINKS'>>(
+expectType<ContentfulClientApi<'WITHOUT_UNRESOLVABLE_LINKS'>>(
   createClient(CLIENT_OPTIONS).withoutUnresolvableLinks
 )
 expectNotAssignable<{ withoutUnresolvableLinks: any }>(
@@ -41,7 +42,7 @@ expectNotAssignable<{ withoutLinkResolution: any }>(
   createClient(CLIENT_OPTIONS).withoutUnresolvableLinks
 )
 
-expectType<Client<'WITHOUT_UNRESOLVABLE_LINKS' | 'WITH_ALL_LOCALES'>>(
+expectType<ContentfulClientApi<'WITHOUT_UNRESOLVABLE_LINKS' | 'WITH_ALL_LOCALES'>>(
   createClient(CLIENT_OPTIONS).withoutUnresolvableLinks.withAllLocales
 )
 expectNotAssignable<{ withoutUnresolvableLinks: any }>(
@@ -54,13 +55,13 @@ expectNotAssignable<{ withoutLinkResolution: any }>(
   createClient(CLIENT_OPTIONS).withoutUnresolvableLinks.withAllLocales
 )
 
-expectType<Client<'WITH_ALL_LOCALES'>>(createClient(CLIENT_OPTIONS).withAllLocales)
+expectType<ContentfulClientApi<'WITH_ALL_LOCALES'>>(createClient(CLIENT_OPTIONS).withAllLocales)
 expectNotAssignable<{ withAllLocales: any }>(createClient(CLIENT_OPTIONS).withAllLocales)
 
-expectType<Client<'WITH_ALL_LOCALES' | 'WITHOUT_LINK_RESOLUTION'>>(
+expectType<ContentfulClientApi<'WITH_ALL_LOCALES' | 'WITHOUT_LINK_RESOLUTION'>>(
   createClient(CLIENT_OPTIONS).withAllLocales.withoutLinkResolution
 )
 
-expectType<Client<'WITH_ALL_LOCALES' | 'WITHOUT_UNRESOLVABLE_LINKS'>>(
+expectType<ContentfulClientApi<'WITH_ALL_LOCALES' | 'WITHOUT_UNRESOLVABLE_LINKS'>>(
   createClient(CLIENT_OPTIONS).withAllLocales.withoutUnresolvableLinks
 )
